add DashboardHome and moved AccountSetting to dashboard folder. Updated navigations in dashboard.
