+++ src/screens/standard/dashboard/DashboardHome.js
@@ -1,0 +1,23 @@
+import React from 'react'
+import { Text, View } from "react-native";
+import { connect } from 'react-redux'
+
+export const DashboardHome = () => {
+  return (
+    <>
+     <View>
+       <Text>DashboardHome</Text>
+      </View> 
+    </>
+  )
+}
+
+const mapStateToProps = (state) => ({
+  
+})
+
+const mapDispatchToProps = {
+  
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(DashboardHome)
