+++ src/svg_assets/SvgEyeballPie.js
